+++ packages/plugin-core/package.json
@@ -1371,11 +1371,13 @@
     "mocha": "^7.1.2",
     "node-loader": "^2.0.0",
     "ovsx": "^0.2.1",
+    "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "source-map-loader": "^3.0.0",
     "terser-webpack-plugin": "^5.2.4",
     "ts-loader": "^8.0.0",
     "ts-node": "^8.10.2",
+    "tsyringe": "^4.7.0",
     "vsce": "^1.100.0",
     "vscode-test": "^1.3.0",
     "webpack": "^5.56.0",
