+++ packages/plugin-core/src/web/test/suite/extension.test.ts
@@ -4,13 +4,12 @@ import * as assert from "assert";
 // as well as import your extension to test it
 import * as vscode from "vscode";
 // import * as myExtension from '../../extension';
-import { controller } from "@dendronhq/plugin-common";
 
 suite("Web Extension Test Suite", () => {
   vscode.window.showInformationMessage("Start all tests.");
 
   test("Sample test", () => {
-    controller();
+    // controller();
     assert.strictEqual(-1, [1, 2, 3].indexOf(5));
     assert.strictEqual(-1, [1, 2, 3].indexOf(0));
   });
